Fix: WebDAV server not starting
Seafile uses it's own version of wsgidav WebDAV server and started by
the seafile controller when `./seafile.sh start` is executed. Problem is
that it calls python3 to do it, but it doesn't know where modified
version of wsgidav located. Since `seafile.sh` don't set `PYTHONPATH`
environment variable, unlike `seahub.sh` which sets it up.

The other problem is you also need to install python requirements for
wsgidav but you don't know which, since they didn't included in the
tarball for the Seafile. There are also no mention of this in Seafile
documentation about WebDAV, or that you need to use custom version of
wsgidav. You can find requirements file here https://github.com/haiwen/seafdav
but it's frustrating that you don't know about this, this is not
mentioned anywhere and you have to find out this by yourself.